refactor: the app owns the engine process, not a window (#3945 PR #1)
The engine was started from a WindowGroup's `.task` (FicheroApp.swift), so its
lifecycle was really owned by whichever window opened first — and
`shouldReuseExistingConnection` existed only to stop every subsequent window/tab
re-running connect+auth (#3394/#3407). Opening a window is not an engine event.

This is the PROCESS axis of the app-owns-engine design (#3947): the app owns one
engine process; each library owns its connection (the connection-set axis lands
in a later slice). Behavior-neutral by construction.

- New `EngineLifecycleController` (@MainActor @Observable): owns the process half
  (EmbeddedBackendService) + the connection half (AppState) + LibraryManager, with
  app-scoped start()/retry()/stop(). `connect()` is a faithful relocation of
  FicheroApp.connectBackend; an `isConnecting` guard replaces the per-window reuse
  guard now that the trigger is one app-scoped start(), not a scene .task.
- `FicheroAppDelegate` is now @MainActor and owns the controller: starts it in
  `applicationDidFinishLaunching`, stops it in `applicationWillTerminate` — the
  engine's async lifecycle on the one app-level hook, independent of any window.
- The three App-struct @State (backendService/appState/libraryManager) become
  computed accessors reading `appDelegate.controller.*`, so the whole App body and
  every .environment injection is unchanged. @Observable is tracked wherever a
  property is read; @State was only ownership, which now belongs to the controller.
- Deleted the WindowGroup .task spawn and the relocated connectBackend/
  showBackendError; onRetry -> appDelegate.controller.retry().

Closes #3973 and #3362/#3394/#3407 by construction (no window can trigger connect),
and unblocks #3975 (redundant-readiness lives in the connect half re-homed here).
`shouldReuseExistingConnection` is left in place (now unreachable from the app) so
its test file isn't orphaned — a follow-up removes it with its tests.

Verified: Dev Embedded BUILD SUCCEEDED (compile + link + embed + sign); swiftlint
clean. Full Swift test suite is the merge gate.
